Fix directory renaming.
