+++ src/soep_preparation/clean_variables/bioedu.py
@@ -2,7 +2,7 @@
 
 import pandas as pd
 
-from soep_preparation.clean_existing_variables import month_mapping
+from soep_preparation.clean_variables import month_mapping
 from soep_preparation.utilities.data_manipulator import (
     apply_smallest_int_dtype,
     object_to_int_categorical,
