+++ src/soep_preparation/clean_variables/ppathl.py
@@ -2,7 +2,7 @@
 
 import pandas as pd
 
-from soep_preparation.clean_existing_variables import month_mapping
+from soep_preparation.clean_variables import month_mapping
 from soep_preparation.utilities.data_manipulator import (
     apply_smallest_float_dtype,
     apply_smallest_int_dtype,
