+++ src/soep_preparation/clean_variables/task.py
@@ -29,7 +29,7 @@ for data_file_name, data_file_catalog in DATA_CATALOGS["data_files"].items():
         raw_data: Annotated[Path, data_file_catalog["raw"]],
         module_path: Annotated[
             Path,
-            SRC / "clean_existing_variables" / f"{data_file_name}.py",
+            SRC / "clean_variables" / f"{data_file_name}.py",
         ],
     ) -> Annotated[pd.DataFrame, data_file_catalog["cleaned"]]:
         """Cleans variables of a data file using the corresponding cleaning module.
