+++ src/soep_preparation/config.py
@@ -28,7 +28,7 @@ else:
 
 
 DATA_FILE_NAMES = get_stems_if_corresponding_raw_data_file_exists(
-    directory=SRC / "clean_existing_variables"
+    directory=SRC / "clean_variables"
 )
 
 DATA_CATALOGS = {
